[one-cmds] Relocate python lib
This commit relocates python lib.

ONE-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>
