+++ compiler/one-cmds/one-import-onnx
@@ -35,7 +35,7 @@ try:
 except ImportError:
     _onnx_legalizer_enabled = False
 
-import python.make_cmd as _make_cmd
+import onelib.make_cmd as _make_cmd
 import utils as _utils
 
 # TODO Find better way to suppress trackback on error
