+++ compiler/one-cmds/one-import-tf
@@ -25,7 +25,7 @@ import subprocess
 import sys
 import tempfile
 
-import python.make_cmd as _make_cmd
+import onelib.make_cmd as _make_cmd
 import utils as _utils
